requirements.txt
attrs~=23.2.0
prettyprinter~=0.18.0
factory-boy~=3.3.0